Pull out NavBarCartButton and wire to CartViewState
